requirements.txt
fastapi>=0.115.6
litellm==1.48.0
livekit_api==0.8.0
python-dotenv==1.0.1
uvicorn==0.30.6
alembic==1.13.3
sqladmin==0.19.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
asyncpg==0.30.0
ruff==0.8.6
deepgram-sdk==3.8.0
pydantic_settings==2.7.1
mcp==1.6.0
pipecat-ai[cartesia,daily,openai,silero,webrtc,deepgram,websocket]==0.0.65
pipecat-ai-small-webrtc-prebuilt==0.5.0
av>=14.0.0
websockets>=13.0.1

openai-agents==0.0.11
pydantic-ai[logfire]==0.0.51
